WIP: Add ?Sized bounds to all (?) appropriate locations
I have added ?Sized to pretty much all T that is garbage collected and
that needs it. However, there are a few structs defined in a rental!
macro that are causing the build to break due to trasmutation rules.
Enforcing a Sized bound causes the GcRef type to require a Sized T,
which doesn't seem right.