AtomicGc design and implementation (#48)
* WIP: Initial draft of AtomicGc work

Adds Atomic support to shredder! The atomic operations should be
fairly fast, but obviously need to do some Gc related book
keeping. Also they block if collection is happening, which may be
inevitible. Perhaps we could lift this limitation for reads?

I think this is the right design--I've thought a lot about other
ways to do this and this was the only fruitful one.

Still missing docs and tests, but wanted to put this up to solicit
feedback.

* Add documentation and tests for `AtomicGc`

Also contains bug fixes around the handling of the `Arc` ptr in
the atomic gc