Implement DerefGc, along with major refactor of the trait system
Kinda too big to summarize. Basically there are new traits
representing whether data can be dropped in the background, as well
as whether it can be deref'd in a DerefGc

Then I redid the derive to support these new traits. Also reorg'd
the way traits are implemented for std types