Use no-std-compat to transition to no-std
This commit uses the no-std-compat crate to enable use of shredder in a
no-std environment. This comes with a few caveats when using a no-std
environment:

- Debug and error print statements are replaced with no-ops
- Locks (Mutexes and RwLocks) use the spin crate
- Hash tables use the hashbrown crate, which is _not_ HashDoS resistant.

I don't believe that any of these constitute a serious problem for
shredder. Shredder can be used in this way by disabling the 'std' feature.